更新特征评估和生成逻辑，核心指标由R²改为MAE，优化模型性能评估流程。同时，新增MVSE特征生成器，增强特征工程能力。调整Sonata实验模块以适应新的评估标准，提升代码可读性和可维护性。
