+++ results/sonata_20250616_131742/iterative_contribution.csv
@@ -1,0 +1,8 @@
+step,feature_name,marginal_contribution,cumulative_performance
+2,rolling_features_macro_set_on_temp,0.020503232365448598,1.779691155857507
+6,"temp_rolling_std_14, temp_rolling_mean_7_x_dayofweek",0.01742187288013408,1.7483971453162523
+4,"temp_lag_1_rolling_std_14, embed_0_win365",0.011781517804577701,1.798273676864539
+3,"temp_lag_2_x_dayofweek, embed_0_win90",0.006801003202454359,1.7864921590599614
+7,"temp_rolling_mean_30, temp_lag_365",-0.002719121559498383,1.745678023756754
+5,temp_rolling_mean_7,-0.06729840442842083,1.7309752724361183
+1,"time_features_macro_set, lag_features_macro_set_on_temp",-1.6396058464460166,1.7591879234920584
