+++ results/sonata_20250616_131742/permutation_importance.csv
@@ -1,0 +1,11 @@
+,importance_mean,importance_std
+temp_lag_1,1.1516634450144663,0.0
+month,0.267461745250825,0.0
+temp_lag_2,0.11959742535717943,0.0
+temp_rolling_mean_7,0.054512069345962644,0.0
+temp_lag_3,0.010813561876897415,0.0
+dayofweek,0.00763327340411446,0.0
+year,0.00701599226101135,0.0
+day,-0.00039197379735389504,0.0
+temp_lag_14,-0.014416778251537421,0.0
+temp_lag_7,-0.016430636385157626,0.0
